[GEOT-5819] ExceptionDetails top-level, javadoc
